Turn big_circle.dart into a test
Fixes #829
